Refactor(src): move utils modules to `src`
